Create playwight helpers
