+++ e2e/src/helpers.ts
@@ -1,0 +1,54 @@
+import type { Page, Browser } from "@playwright/test";
+import { generateAuthenticationToken } from "./context";
+
+// a collection of useful playwright actions
+
+export async function createAuthenticatedSession({
+  browser,
+  userId,
+}: {
+  browser: Browser;
+  userId: number;
+}): Page {
+  const browserContext = await browser.newContext();
+  const page = await browserContext.newPage();
+  const token = generateAuthenticationToken(`${userId}`);
+  await browserContext.addCookies([
+    {
+      name: "jwt",
+      domain: "localhost",
+      path: "/",
+      value: token,
+    },
+  ]);
+  return page;
+}
+
+export async function getAdminPage({
+  browser,
+  userId,
+}: {
+  browser: Browser;
+  userId: number;
+}): Page {
+  const page = await createAuthenticatedSession({ browser, userId });
+  await page.goto("/");
+  await page.waitForResponse((response) => {
+    return response.url().includes("/graphql");
+  });
+  return page;
+}
+
+export async function getTeamPage({
+  browser,
+  userId,
+  teamName,
+}: {
+  browser: Browser;
+  userId: number;
+  teamName: string;
+}): Page {
+  const page = await getAdminPage({ browser, userId });
+  await page.locator("h2", { hasText: teamName }).click();
+  return page;
+}
